refactor(actor): own capability schemas
